+++ src/tasks/doctor.js
@@ -12,6 +12,10 @@ module.exports = function(config, callback) {
 
     gutil.log(doctor.toString());
 
-    callback();
+    if (gutil.env['self-check']) {
+        Doctor.asyncSelfCheck({dir: dir, callback: callback, logger: gutil.log});
+    } else {
+        callback();
+    }
 
 };